backend/requirements.txt
# Core
fastapi==0.118.0
uvicorn==0.37.0
pydantic==2.11.9
pydantic-settings==2.6.1

# Database
psycopg2-binary==2.9.9

# Environment
python-dotenv==1.1.1